+++ script/helpers/generateUniqueId.js
@@ -1,0 +1,3 @@
+export function generateUniqueId() {
+    return new Date().getTime();
+}
