+++ script/helpers/sanitizeInput.js
@@ -1,0 +1,5 @@
+export function sanitizeInput(input) {
+    const maxLength = 100;
+    const sanitizedInput = input.replace(/<\/?[^>]+(>|$)/g, "");
+    return sanitizedInput.substring(0, maxLength);
+}
